requirements.txt
streamlit>=1.28.0
numpy>=1.26.0
pandas>=2.1.0
matplotlib>=3.8.0
plotly>=5.17.0
openpyxl>=3.1.0
python-dateutil>=2.8.2
